Reuse hunter display stat modification
