+++ ui/hunter/survival/sim.ts
@@ -25,6 +25,7 @@ import { HunterStingType, SurvivalHunter_Rotation } from '../../core/proto/hunte
 import * as AplUtils from '../../core/proto_utils/apl_utils';
 import { Stats } from '../../core/proto_utils/stats';
 import * as HunterInputs from '../inputs';
+import { sharedHunterDisplayStatsModifiers } from '../shared';
 import * as SVInputs from './inputs';
 import * as Presets from './presets';
 
@@ -52,25 +53,7 @@ const SPEC_CONFIG = registerSpecConfig(Spec.SpecSurvivalHunter, {
 		Stat.StatMastery,
 	],
 	modifyDisplayStats: (player: Player<Spec.SpecSurvivalHunter>) => {
-		let stats = new Stats();
-		//stats = stats.addStat(Stat.StatMeleeCrit, player.getTalents().lethalShots * 1 * Mechanics.MELEE_CRIT_RATING_PER_CRIT_CHANCE);
-
-		const rangedWeapon = player.getEquippedItem(ItemSlot.ItemSlotRanged);
-		if (rangedWeapon?.enchant?.effectId == 3608) {
-			stats = stats.addStat(Stat.StatMeleeCrit, 40);
-		}
-		if (rangedWeapon?.enchant?.effectId == 4176) {
-			stats = stats.addStat(Stat.StatMeleeHit, 88);
-		}
-		if (player.getRace() == Race.RaceDwarf && rangedWeapon?.item.rangedWeaponType == RangedWeaponType.RangedWeaponTypeGun) {
-			stats = stats.addStat(Stat.StatMeleeCrit, 1 * Mechanics.MELEE_CRIT_RATING_PER_CRIT_CHANCE);
-		}
-		if (player.getRace() == Race.RaceTroll && rangedWeapon?.item.rangedWeaponType == RangedWeaponType.RangedWeaponTypeBow) {
-			stats = stats.addStat(Stat.StatMeleeCrit, 1 * Mechanics.MELEE_CRIT_RATING_PER_CRIT_CHANCE);
-		}
-		return {
-			talents: stats,
-		};
+		return sharedHunterDisplayStatsModifiers(player);
 	},
 
 	defaults: {
